# Speed up boost vector iterators