--- a/rdkit/__init__.py
+++ b/rdkit/__init__.py
@@ -34,3 +34,33 @@ logger.propagate = False
 
 # Uncomment this to use Python logging by default:
 # rdBase.LogToPythonLogger()
+
+#
+#  Boost vector iterators are VERY slow, but indexing is quite fast
+#   this wraps the iterators into a much faster version
+#
+
+class VectIter:
+    def __init__(self, vect):
+        self.vect = vect
+        self.l = len(vect)
+        self.pos = -1
+
+    def __iter__(self):
+        return self
+    
+    def __next__(self):
+        self.pos += 1
+        if self.pos < self.l:
+            return self.vect[self.pos]
+        else:
+            raise StopIteration()
+
+def __vect__iter__(vect):
+    return VectIter(vect)
+
+for name, object in vars(rdBase).items():
+  if name.startswith("_list") or name.startswith("_vect"):
+    if hasattr(object, "__iter__"):
+      object.__iter__ = __vect__iter__
+