Fixes #7639 

Boost Iterators are amazingly slow for iteration, but accessing by index is very fast.  This converts the __iter__ protocol to use the __getitem__ protocol and results in a > 300x speedup for iteration.