@bp-kelley 2024-08-20
I've added the test note book to:

Code/GraphMol/ChemReactions/testData/GetPosition.ipynb 

Here are the tests

```
from rdkit import Chem
m = Chem.MolFromSmiles("CCCCCCCCCCCC")
ps = Chem.FindAllPathsOfLengthN(m, 3)
%timeit [list(x) for x in ps]
```

```
Before: 5.04 ms ± 24.9 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
after: 24.1 μs ± 53.9 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```

```
from rdkit import Chem
from rdkit.Chem import AllChem
from rdkit.Chem import rdChemReactions

for N in range(3, 40):
    carbons = [f"[C:{i}]" for i in range(1,N+1)]
    rxn = f"{'.'.join(carbons)}>>{''.join(carbons)}"
    rxn = AllChem.ReactionFromSmarts(rxn)
    mol = Chem.MolFromSmiles("C")
    bbs = []
    for i in range(N):
        bbs.append( [mol] * 5 )
        
    lib = rdChemReactions.EnumerateLibrary(rxn, bbs)
    print(f"{N=}", len(lib.GetPosition()), end=" ")
    %timeit [x for x in lib.GetPosition()]
```

It looks like the bulk of the cost is returning the position as the timings are not dependent on the return size.  On OSX at least, it's about twice as fast.

Before:
```
N=3 3 652 μs ± 465 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
N=4 4 653 μs ± 2.56 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
N=5 5 652 μs ± 1.28 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
N=6 6 652 μs ± 1.28 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
N=7 7 653 μs ± 1.79 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
N=8 8 654 μs ± 1.55 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
N=9 9 655 μs ± 2.49 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
N=10 10 654 μs ± 1.49 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

After:
```
N=3 3 383 μs ± 1.39 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
N=4 4 383 μs ± 879 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
N=5 5 383 μs ± 874 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
N=6 6 383 μs ± 575 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
N=7 7 382 μs ± 668 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
N=8 8 383 μs ± 697 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
N=9 9 383 μs ± 563 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
N=10 10 383 μs ± 840 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
...
```