@greglandrum 2024-08-21
Running on linux (using g++) I see a factor of ~4 speedup on the first case (the one with FindAllPathsOfLengthN) and no change for the second example.