@bp-kelley 2024-08-21
It looks like the vector wrapped in rdChemReactions isn't in rdBase, adding this adds the speedup

```
rdChemReactions.VectSizeT.__iter__ = __vect__iter__
```
Before
```
N=3 3 24.9 μs ± 283 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
N=4 4 24.9 μs ± 161 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
N=5 5 25.2 μs ± 105 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```

After
```
N=3 3 3.64 μs ± 29.4 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
N=4 4 4.17 μs ± 47.3 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
N=5 5 4.78 μs ± 81.8 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```

I'll see what I can do here, but fixing the underlying vector indexing suite would be a generic solution I think.