@bp-kelley 2024-08-21
The newer tests used exactly the same build.

On Wed, Aug 21, 2024 at 9:49 AM Brian Kelley ***@***.***> wrote:

> I wasn't using the same build for testing which was an oversite.  Maybe
> clang got better at optimizing something since the last release.
>
> On Wed, Aug 21, 2024 at 9:35 AM Greg Landrum ***@***.***>
> wrote:
>
>> It looks like the vector wrapped in rdChemReactions isn't in rdBase,
>> adding this adds the speedup
>>
>> If this is true, and it makes sense that it would be, how did you see the
>> factor of two speed improvement that you showed above?
>>
>> I'll see what I can do here, but fixing the underlying vector indexing
>> suite would be a generic solution I think.
>>
>> Agreed, but that is likely to be "deeply non-trivial"
>>
>> —
>> Reply to this email directly, view it on GitHub
>> <https://github.com/rdkit/rdkit/pull/7719#issuecomment-2302077182>, or
>> unsubscribe
>> <https://github.com/notifications/unsubscribe-auth/ACLOFIAH2Y65ZY6GOWM7EDLZSSJQ5AVCNFSM6AAAAABMQSTC7SVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDGMBSGA3TOMJYGI>
>> .
>> You are receiving this because you authored the thread.Message ID:
>> ***@***.***>
>>
>