@bp-kelley 2024-08-23
We use RegisterVector iterators in a lot of places that would need to be monkey patched.  Like GetPositions.  I'll see how hard this is to consolidate, I think most std::vector<number> should be wrapped at the lowest level, i.e. RDBoost.  And we could wrap std::vector<RWMOL_SPTR/ROMOL_SPTR> in rdkit.Chem which should cover most of the major issues.  I'll get this done this weekend.