@greglandrum 2024-08-25
@bp-kelley maybe that can be done in a separate PR? If you remove the notebook from this one, I can go ahead and merge it and we'll have the partial solution out there in the upcoming patch release.