@bp-kelley 2024-08-25
@greglandrum after this is merged I have a quick follow up where I register numeric vectors in RDBase.cpp, this should fix the original issue.