@bp-kelley 2024-09-03
@greglandrum I think removing the notebook was your change request, let me know if there is anything else